Add Synthesizer trait (#55)
Adds the Synthesizer trait. A program may now optionally return a Synthesizer object that is responsible for generating audio.

This complicates the API, but allows the rendering/simulation thread and the audio thread to be more loosely coupled. Before, the rendering and simulation thread contended for a single Mutex protecting the program. Now, the program returns a discrete Synthesizer object that it shares with the rendering thread. Updates to the Synthesizer object to change playing audio can then be made briefly enough to avoid starving the outgoing audio buffer.