Add filter shaders (#101)
Filter shaders are full-screen postprocessing shaders. They are executed in a chain, with the Nth shader reading the output of the N-1th shader. The first shader in the chain reads the output of the main fragment shader, and the output of the last shader is sent to the display.